Nitro: Implement NCER.
An NCER (nitro cell resource) stores cells, small images formed out of
fixed-size objects.  The image data typically comes from an NCGR, where
it may be stored in one of two different ways: either as a one-
dimensional sequence of 8x8 tiles, or as the more familiar two-
dimensional array of pixels. The former is used by the 2D graphics
engine in the DS, and the latter by the 3D engine.

To deal with the first case, the Tiled type is an implementation of
image.Image which handles tiled images.

I've also added a couple demo programs that helped me test the NCER
implementation, and added a little more input validation to the NCGR
and nitro header parsing functions.