nitro: reset acell states when mcell changes
When an animation transitioned from one mcell to another, we would carry
the acell state over. If the new mcell reused acells that had already
been used, they might start in the middle of their animation and then
loop back, causing weird jumps in the animation.  For example, in
Jellicent's frontsprite and backsprite, and in Mewtwo's tail.

I'm not sure what the correct behaviour is here - maybe acell state
needs to be per-mcell instead of being global - but this seems to fix
the examples i know about.